refactor(billing): move the plan-limit formatter out of the pricing card
`formatLimitValue` decides how a plan limit is written for a reader - "Custom",
"Unlimited", "500 MB", "10 GB" - and it lived inside a component, where
product-copy.ts cannot reach it. That component already takes nine names from
product-copy.ts, so importing back would close a cycle, and product-copy.ts is
read by llms[.]txt.ts, seo.ts and seo-registry.ts, none of which render, so
reaching into a .tsx module would pull Radix and lucide into their graphs.

Moved verbatim into a leaf module with no imports of its own, with its spec.
No behavior change: the function is byte-identical and the pricing cards call
it from its new home.
